+++ src/main/groovy/com/bmuschko/gradle/vagrant/validation/BackendProviderAware.groovy
@@ -1,0 +1,5 @@
+package com.bmuschko.gradle.vagrant.validation
+
+interface BackendProviderAware {
+    void setProvider(String provider)
+}