+++ src/main/groovy/org/gradle/api/plugins/vagrant/validation/BackendProviderAware.groovy
@@ -1,5 +1,0 @@
-package org.gradle.api.plugins.vagrant.validation
-
-interface BackendProviderAware {
-    void setProvider(String provider)
-}